2.1pre29
{
    "filename": "links-2.1pre29.tar.gz",
    "md5": "c3373688f163afa1236cbd025635e538",
    "sha1": "16378076d6f463b2ab85efb987726199e82364a7",
    "sha256": "0deb178c7c72b102270c90f0159cbbe9e7de808b3dbd1e11f9f3c4f646c5e8cc"
}

+++ bfu.c
@@ -447,7 +447,7 @@ void menu_func(struct window *win, struct event *ev, int fwd)
 	struct menu *menu = win->data;
 	struct window *w1;
 	menu->win = win;
-	switch (ev->ev) {
+	switch ((int)ev->ev) {
 		case EV_INIT:
 		case EV_RESIZE:
 			get_parent_ptr(win, &menu->xp, &menu->yp);
@@ -676,7 +676,7 @@ void mainmenu_func(struct window *win, struct event *ev, int fwd)
 	int s = 0;
 	struct mainmenu *menu = win->data;
 	menu->win = win;
-	switch(ev->ev) {
+	switch ((int)ev->ev) {
 		case EV_INIT:
 		case EV_RESIZE:
 #ifdef G
@@ -1155,7 +1155,7 @@ void dialog_func(struct window *win, struct event *ev, int fwd)
 		return;
 	}
 	
-	switch (ev->ev) {
+	switch ((int)ev->ev) {
 		case EV_INIT:
 			for (i = 0; i < dlg->n; i++) {
 				struct dialog_item_data *di = &dlg->items[i];
